Modify Error entity class in order to implement error/search endpoint.
The enum fields of this entity have been annotated appropriately to ensure
that they are saved as Strings. This is important because the error / search
endpoint must allow queries to the database to search for error aggregates
having as search criteria the values of the environment and level columns,
which are enums.

In addition, the createdAt field received @Column(updatable = false) annotation
to indicate that it should not be updated. This is important because queries
made to the database by the endpoint error / search assume that this value is
marked at the time of error insertion and is never updated thereafter.