Implements queries to fetch error aggregates through the @Subselect annotation.
This implementation is based on using hibernate's @Subselect annotation to map
an entity representing an aggregate of errors to the SQL query Result set that
returns that aggregate. This method allows you to produce an immutable JPA
entity associated with this query, as if the query result were a table.
The advantage of this is that from then on the entity in question can be used
like any other JPA entity, allowing te creation of a specific JPA spring data
repository for it, the use of query methods to perform selects, etc.
This eliminates several issues encountered during the implementation
of the error aggregate search endpoint.