Add two new pom dependecies: specification arg resolver and gson.
specification arg resolver is used to facilitate the creation of endpoints that
enable applying multiple filters to a given query. The library in question
provides specific annotations for endpoint parameters dedicated to
filtering results. From these annotations specification arg resolver creates
the appropriate Specifications for each filter parameter.
It is used in error aggregate search endpoint.

Gson is useful for serializing and de-serializing json objects.
It is used in the Controller Advicer of ErrorController.